compaction_additional_test: wait until level distribution converges
for a similar reason of the issue addressed by 6690032e6a,
test_lcs_sstable_promotion is also flaky.

in this change, we apply the same cure to this problem: to wait
until the level distribution converges.

despite that the description of the issue does not involve
test_lcs_table_promotion_major_compaction, it is updated as well.
because it would also suffer from the same problem even after the
test is fixed.

Fixes https://github.com/scylladb/scylladb/issues/15714
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>