compaction_additional_test: consider [0,0,44] a well-formed LCS shape
if all sstables are pushed to L3, and the number of sstables is under
the size limit at that level, we should consider this distribution
follows the requirement of LCS. to be more specific,

 L1 = 0, L2 = 0, and L3 = 44,

this is a perfectly fine distribution of sstables from LCS's
perspective. actually, this is the best case of the LCS. as
as the last level is filled, and we don't have any overlapped data
in its lower levels.

the total size of sstables on L3 is 10**3. after compacting all
sstables, L3 is able to hold all data without breaching the
constraints enforced by LCS:

- the space amplification factor is 0, which is under 10%,
- the read amplification is 0. at most a single disk read would suffice
  to serve a certain read query.

so, in this change we both tighten the verification and relatex it.

* ensure the number of sstables on each levels is under the bound.
* do not check the level_n * 10 <= level_[n+1], as this contraint is
  not a part of LCS.
* do not consider the case where all sstables are located in the same
  level a bad distribution.

Fixes #2938
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>