fix(twcs_additional_test): fix default table's ttl
Test didn't wait enough for data to be ttl'ed and it happened that
Scylla was compacting marked sstables along with unexpired sstables.

Added 30s margin so compaction starts when we are sure sstables are
expired. Adjusted default table ttl to 120s so inserted data by c-s is
not expiring before we enable autocompaction.

fixes: https://github.com/scylladb/scylla-dtest/issues/3642