fix(replace_address_test): ignore closed_error for replace with same ip
The flakiness of test_replace_node_same_ip is described in scylladb/scylladb#16713:

"Nodes are dropping all connections to nodes about which they learn of being DOWN,
but this is racing with the test. If the test replacing quickly enough, the leader
might have not dropped the connections yet and could try reusing some of them e.g.
for snapshot transfer RPC.", -- kbr-scylla said.

Just ignore the error in test's teardown.