bootstrap_test: test_smallest_ip_join_late: generalize expected errors
In ca39cafe0340cbfa9b7dc168f8f551526518e1c9 the test was generalized so
that we allow bootstrap to fail during shadow round on one of the nodes,
to prepare for PR scylladb/scylladb#15941 which makes the shadow round
safer.

Unfortunately, this is not enough. There is a race in the test: the two
nodes are started concurrently, and the point at which during the
bootstrap procedure they fail, depends on their relative speed of
booting. Namely `node3` may observe `node2` with `status=UNKNOWN` in
gossiper if `node2` boots relatively slow compared to `node3`, or if
`node2` boots quickly and fails quickly, `node3` may not observe `node2`
at all and fail with a different message.

So this commit further generalizes the test: instead of waiting for a
particular set of messages to appear in the nodes' logs, it simply
expects the process to eventually stop with exit code 1 (indicating any
failure).

`ignore_log_patterns` still needs to be extended with the list of
messages so that the test ignores the ERRORs in nodes' logs. But the
definition of the list, `start_failure_msgs`, can now be simplified,
each message as a separate entry in the list instead of having some of
them concatenated into a single regex pattern.

I have executed the test 20 times with and without the PR to check that
it passes.

(cherry picked from commit 25f52f64872bbd72b739b207468cfe620d924a1f)