database_metrics_test: accept per-service-level metrics for enterprise
Prior to scylladb/scylla-enterprise@05b79935f6f42769d230d3191ed8f4b2278e88f4,
the reader_concurrency_semaphore metrics that are collected per
service level are aggregated in the server and presented as the
'user' class, same as OSS. After that commit, those metrics
are exposed for every service level individually. This is better as it
allows measuring the impact of different workloads, but breaks the
metrics test.

Fix it by looking for 'sl:default' class (the service level used when
none are configured), and falling back to 'user' if not found. This works
with OSS, pre-change enterprise, and post-change enterprise.