replace_address_test: wait for table to be created before stopping node
61518e3c416ff8f8be5ba598ea99159323a8b8eb change the way the background
thread are running, and surfaced the issue the we are not waiting
enough time for c-s to be running, before stopping node3.

this fix adds wait for the table to be create and can be read from
before moving on to the next step of the test.

Fix: scylladb/scylladb#17580
(cherry picked from commit 5556e9b3ac9090686414628c77afc6b87196fc94)