nodetool_additional_test.py: skip random corruption tests
These tests are testing a functionality which doesn't exist -- scrub
recovering arbitrary corruptions. Scrub was never designed to do that and
it fails in unexpected ways in this test from time-to-time.
We plan to make scrub be able to *detect* (but not *recover*) arbitrary
corruption, but that is not implemented yet.
TODO: re-enable and refactor these tests once the above is implemented.

Refs: scylladb/scylladb#15693
Fixes: scylladb/scylladb#15457
Fixes: scylladb/scylladb#15195