tools/data: lower default concurrency to 20
we have multiple calls to `execute_concurrent_with_args`
which doens't set concurrency, so the default is 100

for some tests it's not needed, and can be too high.
we now would default it to 20, and each test can now
change if needed.

Ref: scylladb/scylladb#15715
Ref: scylladb/scylladb#18001
Ref: scylladb/scylla-dtest#4115
(cherry picked from commit 3fba3a59933845739d25dcc66a363069cdf58e10)