repair_additional_test.py: test_repair_compacts_data: fix tx row estimate
The test writes one partition with 10 rows and one partition tombstone
covering the row, then checks that with compaction enabled on the
repair, repair will send over just the tombstone, but not the rows
covered by it.
Then the test writes another similar partition, disables compaction on
repair, and expects all 10 rows to be sent over by repair. This
expectation is incorrect: once compaction on repair is disabled, all 10
rows from the first partition, as well as the 10 rows + 1 tombstone from
the second one are transmitted. The only reason the test was passing is
that the live-update of the config which disables compaction was broken
and thus the second repair also transmitted 1 tombstone only (which is
less then the expected max 10 tx).

Discovered in the CI of the PR which fixes the live-update:
https://github.com/scylladb/scylladb/pull/18705

(cherry picked from commit 87e1106448f3edc7b08aac84fc0bcb0523ad4822)