internode_ssl_test: test_listen_ports_conf: adjust to current behavior
The disable_value=None case was skipped due to
https://github.com/scylladb/scylladb/issues/7500
But the issue is about scylla still listening
on the unencrypted port, which is no longer the case.

The current failures are about scylla listening on
the `ssl_storage_port` with `internode_encryption="all"`
and `disable_value=None`, which are unrelated to the
above issue.

It appears like scylla is bahving as designed
and documented for the config option:
```
The SSL port for encrypted communication. Unused unless enabled in encryption_options.
```

https://opensource.docs.scylladb.com/stable/operating-scylla/security/node-node-encryption.html
should be beefed up about how to configure `storage_port`, and
`ssl_storage_port`, since https://opensource.docs.scylladb.com/stable/operating-scylla/admin.html
does not document them.

That said, using the default when it is not overridden
in scyla.yaml (in the `disable_value=None` case) is the expected behavior.
And disabling the port with `disable_value=0` also works as expected,
so we just need to adjust the test to check that.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>
(cherry picked from commit dd3e1a0044daa4f9fafe22d61c7be7ca0f6e886c)