system_tables_test.py: test_cache_metrics: relax entry count check
The test writes N rows, flushes the memtable, then expects that the
cache has exactly N entries. This is only the case if there is no other
query happening on the system, which of course is not guaranteed. Relax
the check to use >= instead of ==, allowing for internal queries
populating the cache.

Fixes: https://github.com/scylladb/scylladb/issues/17778
(cherry picked from commit 01fce6f43f772c2e3e2fe7fc853957bca02c8334)