dtest_setup: limit the output in case of lots of errors from scylla
during teardown we collect all of the error from scylla log
and raise it as exception, this can fill the output, and cause the
jenkins pluging that retry base on errors from the output to get stuck
for a long duration.

in this change we limit show the 5 first errors from each node
in the exception, the full information would be available on dtest log.

Ref: https://github.com/scylladb/scylla-dtest/issues/4520
(cherry picked from commit bf872566f75a8266b7ac6db669082f20d862204d)