refactor(SPLIT_TIME_TARGET): Reduce target time to 180 for gating and release
In order to complete the gating and release testing quicker we can reduce
the splot target time to 180 which will target to complete these test jobs
in about 30m. This is an addition to the optimizations of the parallel
running and removing the stages of 'heavy' and 'long'.

Another related optimization is in scylla-pkg to use faster insatnces
as part of dtest ASG.

(cherry picked from commit 5061e33680bcc436ff89a13e06776b247194f7c8)