dtest_setup.py: fix loop variable leakage
The value of `name` is overwritten with the name of the last node.
The first time the code is ran, it will create a broken symlink to nodeX.
The seccond time it will not remove it, beucase `os.path.exists` returns False
for broken symlinks, and `os.symlink` will fail because the file exists.

(cherry picked from commit 44d7e23c7dabbde1eb4248776253d33c22bea444)